Resolve  #1957: Create base class for KeyValueCursor and Builder, use with IndexPrefetchRangeKeyValueCursor
